Time: Break up CLI into flags, output modules
